flutter_calendar_carousel.dart: refactor choosing dayStyle
From now `nextTextStyle` and `prevDaysTextStyle` more important then `weekendDaysTextStyle`. As we have no special style for prev weekends and next weekends, this is he most simple way how we can hide prev and next month dates (make it white). Otherwise even if that white - that shows red as a weekends.

In the future will be better add setting `showNextMonthDays` and `showPrevMonthDays`.

Also refactor a little method of choosing the current style.